refactor: specific folders added
