+++ src/Components/HeroSection/HeroSection.js
@@ -2,6 +2,7 @@ import React from 'react'
 import "./HeroSection.css"
 import {Route, Switch} from "react-router-dom"
 import { History, Home, LikedVideos, Playlist, VideoDetail, WatchLater } from '../../Pages'
+import { Categories } from '../../Pages/Categories/Categories'
 
 export const HeroSection = () => {
     return (
@@ -13,6 +14,7 @@ export const HeroSection = () => {
                     <Route path="/liked-videos" component={LikedVideos}/>
                     <Route path="/watch-later" component={WatchLater}/>
                     <Route path="/video-detail/:videoId" component={VideoDetail}/>
+                    <Route path="/categories" component={Categories}/>
                 </Switch>
         </div>
     )
