+++ src/Pages/Categories/Categories.css
@@ -1,0 +1,3 @@
+.categoryContainer{
+    margin: 0 0 0 1rem;
+}
