+++ src/Pages/VideoDetail/VideoDetail.css
@@ -1,7 +1,7 @@
 .videoDetailContainer{
     padding: 3rem 0 0 1rem;
     display: grid;
-    grid-template-columns: 3fr 1fr;
+    grid-template-columns: 2fr 1fr;
     grid-gap: 1rem;
     min-height: 90vh;
 }
@@ -25,6 +25,14 @@
 .videoLike{
     display: flex;
     flex-direction: row;
+}
+.videoChannelName{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+.videoChannelName p{
+    margin-left: 1rem;
 }
 .input{
     width: 18rem;
