+++ src/Pages/WatchLater/WatchLater.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
-import { VideoItem } from '../Components';
-import { useStore } from '../Store/storeContext'
+import { VideoItem } from '../../Components'
+import { useStore } from '../../Store/storeContext'
 import "./WatchLater.css"
 
 export const WatchLater = () =>{
