+++ src/Pages/index.js
@@ -1,6 +1,6 @@
-export {History} from "./History"
-export {Home} from "./Home"
-export {LikedVideos} from "./LikedVideos"
-export {Playlist} from "./Playlist"
-export {WatchLater} from "./WatchLater"
-export {VideoDetail} from "./VideoDetail"
+export {History} from "./History/History"
+export {Home} from "./Home/Home"
+export {LikedVideos} from "./LikedVideos/LikedVideos"
+export {Playlist} from "./PlayList/Playlist"
+export {WatchLater} from "./WatchLater/WatchLater"
+export {VideoDetail} from "./VideoDetail/VideoDetail"
